refactor: SuccessCode, ErrorCode 패키지 분리
