+++ src/main/java/org/sopt/global/exception/SuccessCode/SuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.global.exception.constant;
+package org.sopt.global.exception.SuccessCode;
 
 public interface SuccessCode {
     int getStatus();
